regresion-server/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
scikit-learn
python-dotenv